requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.1
python-dotenv==1.0.1
httpx==0.28.1
numpy==1.26.4
pandas==2.1.4
scikit-learn==1.4.0
shap==0.46.0